refactor: remove dead MCP tool-input schemas and tests
These Zod schemas, type aliases, and re-export wrappers were remnants
of the former MCP tool surface (codex/claude/coral-agent). MCP was
fully replaced by CLI-only invocation but the validation layer remained
with zero runtime imports.

Relocate the mapTurnStartParams effort-mapping tests to a dedicated
request-mapping.test.ts before deleting the host file.
